+++ tests/dataframe.test.js
@@ -773,3 +773,129 @@ test("DataFrame can be shuffled", assert => {
         "randomly shuffled and get the same length when having one row."
     );
 });
+
+test("DataFrame can be subsetted", assert => {
+    const df = new DataFrame([...Array(20).keys()].map(row => [row]), ["c1"]);
+
+    assert.deepEqual(
+        df.slice(0, 10).toCollection(),
+        [
+            { c1: 0 },
+            { c1: 1 },
+            { c1: 2 },
+            { c1: 3 },
+            { c1: 4 },
+            { c1: 5 },
+            { c1: 6 },
+            { c1: 7 },
+            { c1: 8 },
+            { c1: 9 }
+        ],
+        "slicing the dataframe between 0 and 10 index."
+    );
+
+    assert.deepEqual(
+        df.slice(0, 10).toCollection(),
+        [
+            { c1: 0 },
+            { c1: 1 },
+            { c1: 2 },
+            { c1: 3 },
+            { c1: 4 },
+            { c1: 5 },
+            { c1: 6 },
+            { c1: 7 },
+            { c1: 8 },
+            { c1: 9 }
+        ],
+        "slicing the dataframe out of index."
+    );
+
+    assert.deepEqual(
+        df.head().toCollection(),
+        [
+            { c1: 0 },
+            { c1: 1 },
+            { c1: 2 },
+            { c1: 3 },
+            { c1: 4 },
+            { c1: 5 },
+            { c1: 6 },
+            { c1: 7 },
+            { c1: 8 },
+            { c1: 9 }
+        ],
+        "getting a new dataframe with the first 10 elements."
+    );
+
+    assert.deepEqual(
+        df.head(2).toCollection(),
+        [{ c1: 0 }, { c1: 1 }],
+        "getting a new dataframe with the first 2 elements."
+    );
+
+    assert.deepEqual(
+        df.tail().toCollection(),
+        [
+            { c1: 10 },
+            { c1: 11 },
+            { c1: 12 },
+            { c1: 13 },
+            { c1: 14 },
+            { c1: 15 },
+            { c1: 16 },
+            { c1: 17 },
+            { c1: 18 },
+            { c1: 19 }
+        ],
+        "getting a new dataframe with the last 10 elements."
+    );
+
+    assert.deepEqual(
+        df.tail(2).toCollection(),
+        [{ c1: 18 }, { c1: 19 }],
+        "getting a new dataframe with the last 2 elements."
+    );
+});
+
+test("DataFrame row", assert => {
+    const df = new DataFrame([...Array(20).keys()].map(row => [row]), ["c1"]);
+
+    assert.deepEqual(
+        df.getRow(2).toDict(),
+        { c1: 2 },
+        "can be accessed by its index."
+    );
+    assert.deepEqual(
+        df.getRow(300),
+        undefined,
+        "can be accessed by its index."
+    );
+
+    assert.deepEqual(
+        df.setRow(2, row => row.set("c1", 656)).toCollection(),
+        [
+            { c1: 0 },
+            { c1: 1 },
+            { c1: 656 },
+            { c1: 3 },
+            { c1: 4 },
+            { c1: 5 },
+            { c1: 6 },
+            { c1: 7 },
+            { c1: 8 },
+            { c1: 9 },
+            { c1: 10 },
+            { c1: 11 },
+            { c1: 12 },
+            { c1: 13 },
+            { c1: 14 },
+            { c1: 15 },
+            { c1: 16 },
+            { c1: 17 },
+            { c1: 18 },
+            { c1: 19 }
+        ],
+        "can be modified by its index."
+    );
+});